requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
google-generativeai>=0.8.5
requests>=2.31.0